send booking invoices through stripe

+++ app/workers/send_booking_invoice.rb
@@ -9,6 +9,7 @@ class SendBookingInvoice
     mg_client = Mailgun::Client.new ENV['MAILGUN_KEY']
 
     invoice = Invoice.find(args['invoice_id'])
+    booking = invoice.booking
     current_user = User.find(args['user_id'])
     settings = Setting.first
     errors = []
@@ -45,28 +46,35 @@ class SendBookingInvoice
 
     # send invoice
     is_test_mode = ENV['TEST_MODE'] == 'true'
-    invoice.export!(pathname)
-    attachments = [File.open("#{pathname}/Invoice #{invoice.number}.pdf", "r"), File.open(Rails.root.join('app', 'workers', 'Film Movement W9.pdf'), "r")]
-    message_params = {
-      from: current_user.email,
-      to: (is_test_mode ? ENV['TEST_MODE_EMAIL'] : args['email']),
-      cc: (is_test_mode ? nil : current_user.email),
-      subject: subject,
-      text: text,
-      attachment: attachments
-    }
-    begin
-      mg_client.send_message 'filmmovement.com', message_params
-      Setting.first.update(next_booking_invoice_number: settings.next_booking_invoice_number + 1) # update next booking invoice number
-    rescue
-      errors << "Unable to send invoice to #{args['email']}"
-    end
 
-    if errors.present?
-      invoice.destroy
-      job.update!({ status: :failed, first_line: "Failed to Send Invoice", errors_text: errors.join("\n") })
-    else
+    if booking.get_use_stripe
+      invoice.create_in_stripe!
+      invoice.email_through_stripe!
       job.update!({ status: :success, first_line: "Invoice Sent Successfully", metadata: { showSuccessMessageModal: true } })
+    else
+      invoice.export!(pathname)
+      attachments = [File.open("#{pathname}/Invoice #{invoice.number}.pdf", "r"), File.open(Rails.root.join('app', 'workers', 'Film Movement W9.pdf'), "r")]
+      message_params = {
+        from: current_user.email,
+        to: (is_test_mode ? ENV['TEST_MODE_EMAIL'] : args['email']),
+        cc: (is_test_mode ? nil : current_user.email),
+        subject: subject,
+        text: text,
+        attachment: attachments
+      }
+      begin
+        mg_client.send_message 'filmmovement.com', message_params
+        Setting.first.update(next_booking_invoice_number: settings.next_booking_invoice_number + 1) # update next booking invoice number
+      rescue
+        errors << "Unable to send invoice to #{args['email']}"
+      end
+
+      if errors.present?
+        invoice.destroy
+        job.update!({ status: :failed, first_line: "Failed to Send Invoice", errors_text: errors.join("\n") })
+      else
+        job.update!({ status: :success, first_line: "Invoice Sent Successfully", metadata: { showSuccessMessageModal: true } })
+      end
     end
   end
 end
